+++ src/main/java/ahmed/foudi/service/interfaces/StageServiceI.java
@@ -5,10 +5,10 @@ import ahmed.foudi.entities.Stage;
 import java.util.List;
 
 public interface StageServiceI {
-    Stage findById(int id);
+    Stage findById(Long id);
     List<Stage> findAll();
     void save(Stage stage);
     void delete(Stage stage);
     void update(Stage stage);
-    void delete(int id);
+    void delete(Long id);
 }
